third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/008-expected.txt

This is a testharness.js-based test.
FAIL Replacement of document prototype object after document.open assert_equals: expected (undefined) undefined but got (number) 1
Harness: the test ran to completion.
